- Code refactoring as per SOLID Principles
