+++ swift/Sources/GildedRose/Items/CustomisedItem.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Manali Mogre on 17/08/2020.
+//
+
+import Foundation
+
+protocol CustomisedItem {
+    var item: Item {get set}
+    func updateItemState()
+}
